Distinguish greedy/non-greedy repetitions
Previously when we follow epsilon transitions on `Split`
instructions, we recurse on the first branch and then
recurse on the second.

In this way all the next PC positions are pushed onto
the next list, so the match is always greedy.

Now for non-greedy repetitions, if we find a match when
recursing on the first branch, we return immediately to
keep conservative.

We stop iterating on the `curr` and `next` lists until there's
no more recorded PCs in the `curr` list.

So for greedy repetitions there will always be something to look
at since we try the repetition first thus it comes before
the non-greedy possibilities on the `next` list. But for
non-greedy repetitions, since we early stop following epsilon
transitions, the other branch, the greedy one, will not be tried
at all thus it won't appear in `next` list.