Avoid cloning at Split instructions
Now we set up a stack to keep track of the following of
epsilon transitions. An action is pushed onto the stack
to restore the captured groups so that we can reuse the
same Thread struct during this procedure. Only before a
thread is inserted into the `next` list we clone the
captured groups.